requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.1
django-jet==1.0.8
django-jet-4==0.0.1
django-js-asset==2.0.0
django-meta==2.2.0
django-nose==1.4.7
djangorestframework==3.14.0
idna==3.4
leaflet==0.0.3
nose==1.3.7
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
urllib3==2.0.2
